split settings tabs into separate files
